--- src/lightning_ml/core/data/datasets/abstract.py
@@ -1,92 +1,0 @@
-"""
-Abstract mix-in and base classes for dataset definitions.
-
-This module provides:
-- InputMixin, TargetMixin, PositiveMixin, NegativeMixin: enforce specific `get_` methods and sample_keys.
-- UnlabelledDatasetBase, LabelledDatasetBase, ContastiveDatasetBase, TripletDatasetBase: abstract dataset base classes combining mix-ins.
-"""
-
-from typing import Any
-from collections.abc import Sequence
-
-# Import ``BaseDataset`` directly from the core package to avoid relying on a
-# non-existent ``data.core`` module and to prevent circular imports when the
-# top-level package is initialised.
-from lightning_ml.core.data.dataset import BaseDataset
-
-__all__ = [
-    "InputMixin",
-    "TargetMixin",
-    "PositiveMixin",
-    "NegativeMixin",
-    "UnlabelledDatasetBase",
-    "LabelledDatasetBase",
-    "ContastiveDatasetBase",
-    "TripletDatasetBase",
-]
-
-
-class InputMixin(BaseDataset):
-    """Dataset mix-in that mandates a `get_input` method, the output of which
-    will be added to the `__getitem__` ouput dict of all subclasses with key
-    "input".
-    """
-
-    sample_keys = ["input"]
-
-    @property
-    def in_shape(self):
-        return self.get_input(0).shape
-
-
-class TargetMixin(BaseDataset):
-    """Dataset mix-in that mandates a `get_target` method."""
-
-    sample_keys = ["target"]
-
-
-class PositiveMixin(BaseDataset):
-    """Dataset mix-in that mandates a `get_positive` method."""
-
-    sample_keys = ["positive"]
-
-
-class NegativeMixin(BaseDataset):
-    """Dataset mix-in that mandates a `get_negative` method."""
-
-    sample_keys = ["negative"]
-
-
-# Abstract dataset bases
-
-
-class UnlabelledDatasetBase(InputMixin):
-    """
-    Abstract base for unlabeled datasets.
-
-    Combines mix-ins to enforce retrieval of input samples only.
-    """
-
-
-class LabelledDatasetBase(InputMixin, TargetMixin):
-    """
-    Abstract base for labeled datasets.
-
-    Combines mix-ins to enforce retrieval of input and target samples.
-    """
-
-
-class ContastiveDatasetBase(InputMixin, PositiveMixin):
-    """
-    Abstract base for contrastive datasets.
-
-    Combines mix-ins to enforce retrieval of input and positive samples.
-    """
-
-
-class TripletDatasetBase(ContastiveDatasetBase, NegativeMixin):
-    """
-    Abstract base for triplet datasets.
-
-    Combines mix-ins to enforce retrieval of input, positive, and negative samples.
-    """
