+++ src/lightning_ml/core/data/loaders/__init__.py
@@ -7,10 +7,9 @@ from lightning_ml.core.utils.registry import get_registry
 
 # Define the registry before importing submodules that register with it to
 # avoid import cycles during package initialisation.
-REGISTRY = get_registry(Registries.DATASET)
+REGISTRY = get_registry(Registries.LOADER)
 
 # Import dataset implementations which will register themselves with
 # ``DATASET_REG`` on import.
-from .contrastive import *  # noqa: F401,F403
-from .labelled import *  # noqa: F401,F403
-from .unlabelled import *  # noqa: F401,F403
+from .csv import *  # noqa: F401,F403
+from .folder import *  # noqa: F401,F403
