--- src/lightning_ml/datasets/__init__.py
@@ -1,3 +1,0 @@
-"""Alias module exposing datasets at :mod:`lightning_ml.datasets`."""
-
-from ..core.data.datasets import *  # noqa: F401,F403
